+++ routes/api.php
@@ -5,6 +5,11 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 
 
+// forrequest para validaciones de la api
+// controlar el borrado de eventos (usuario y ponente al la inscripcion) //hecho comprobar
+// proteger peticiones a la api
+// las fotos (material de belen)
+
 
 use App\Http\Controllers\PonenteController;
 use App\Http\Controllers\EventoController;
@@ -17,24 +22,22 @@ use Laravel\Sanctum\Sanctum;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Controllers\PayPalController;
 
-// 🔹 Autenticación
-Route::get('/user', function () {
-    return response()->json(Auth::user());
-});
 
 
 Route::middleware('auth:sanctum')->post('/paypal/create-order', [PayPalController::class, 'createOrder']);
 Route::middleware('auth:sanctum')->post('/paypal/capture/{orderID}', [PayPalController::class, 'captureOrder']);
 
-
+Route::middleware('auth:sanctum')->post('/procesarPago', [PagoController::class, 'procesarPago']);
 
 Route::post('/login', [AuthController::class, 'login']);
 Route::post('/register', [AuthController::class, 'register']);
+
+
 Route::get('/ponentes', [PonenteController::class, 'index_api']);
-Route::get('/ponentes/{id}', [PonenteController::class, 'show_api']);
-Route::post('/ponentes', [PonenteController::class, 'create_api']);
-Route::put('/ponentes/{id}', [PonenteController::class, 'update_api']);
-Route::delete('/ponentes/{id}', [PonenteController::class, 'destroy_api']);
+Route::middleware('auth:sanctum')->get('/ponentes/{id}', [PonenteController::class, 'show_api']);
+Route::middleware(['auth:sanctum','admin'])->post('/ponentes', [PonenteController::class, 'create_api']);
+Route::middleware(['auth:sanctum','admin'])->put('/ponentes/{id}', [PonenteController::class, 'update_api']);
+Route::middleware(['auth:sanctum','admin'])->delete('/ponentes/{id}', [PonenteController::class, 'destroy_api']);
 
 
 
@@ -43,17 +46,17 @@ Route::get('/asistencias', [AsistenciaController::class, 'index_api']);
 Route::get('/asistencias/{id}', [AsistenciaController::class, 'show_api']);
 
 // Rutas para el dashboard del usuario
-Route::get('/dashboard', [DashboardController::class, 'index_api']);
+Route::middleware('auth:sanctum')->get('/dashboard', [DashboardController::class, 'index_api']);
 Route::get('/home', [HomeController::class, 'index_api']);
 
 
 
 //eventos
 Route::get('/eventos', [EventoController::class, 'index_api']);
-Route::get('/eventos/{id}', [EventoController::class, 'show_api']);
-Route::post('/eventos', [EventoController::class, 'store_api']);
-Route::put('/eventos/{id}', [EventoController::class, 'update_api']);
-Route::delete('/eventos/{id}', [EventoController::class, 'destroy_api']);
+Route::middleware('auth:sanctum')->get('/eventos/{id}', [EventoController::class, 'show_api']);
+Route::middleware(['auth:sanctum','admin'])->post('/eventos', [EventoController::class, 'store_api']);
+Route::middleware(['auth:sanctum','admin'])->put('/eventos/{id}', [EventoController::class, 'update_api']);
+Route::middleware(['auth:sanctum','admin'])->delete('/eventos/{id}', [EventoController::class, 'destroy_api']);
 
 
 
@@ -65,33 +68,6 @@ Route::middleware('auth:sanctum')->delete('/eventos/{evento}/desapuntarse', [Eve
 // Ruta para obtener las inscripciones del usuario
 Route::middleware('auth:sanctum')->get('/user/inscripciones', [AuthController::class, 'inscripciones']);
 
-
-// 🔹 Rutas protegidas con autenticación
-//Route::middleware(['auth:sanctum'])->group(function () {
-//
-//    // Usuario
-//    Route::get('/user', function (Request $request) {
-//        return $request->user();
-//    });
-//
-//    // 🔹 Ponentes
-////    Route::apiResource('ponentes', PonenteController::class);
-//
-//
-//    // 🔹 Eventos
-//    Route::apiResource('eventos', EventoController::class);
-//    Route::post('/eventos/{evento}/inscribirse', [EventoController::class, 'inscribirse']);
-//    Route::delete('/eventos/{evento}/desapuntarse', [EventoController::class, 'desapuntarse']);
-//
-//
-//    // 🔹 Rutas para administradores (middleware adicional si es necesario)
-//    Route::middleware(['admin'])->prefix('admin')->group(function () {
-////        Route::apiResource('ponentes', PonenteController::class)->except(['index', 'show']);
-//        Route::apiResource('eventos', EventoController::class)->except(['index', 'show']);
-//        Route::apiResource('pagos', PagoController::class)->except(['store']);
-//        Route::get('/asistencias', [AsistenciaController::class, 'index']);
-//        Route::delete('/asistencias/{asistencia}', [AsistenciaController::class, 'destroy']);
-//    });
 
     // 🔹 Cerrar sesión
     Route::post('/logout', [AuthController::class, 'logout']);
